% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{discrete_color}
\alias{discrete_color}
\title{discrete palette for colour aesthetic}
\usage{
discrete_color()
}
\value{
discrete scale
}
\description{
n <= 9: RColorBrewer Set1\cr  
n > 9: hues "I want hue"
}
\examples{
ggplot2::ggplot(data.frame(x=c("A","B","C"),y=1:3), aes(x=x,y=y,colour=x)) + geom_point() + discrete_colour()

}

