% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{discrete_palette}
\alias{discrete_palette}
\title{discrete palette}
\usage{
discrete_palette(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
vector of colors
}
\description{
n <= 9: RColorBrewer Set1\cr 
n > 9: hues "I want hue"
}
