% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{discrete_fill}
\alias{discrete_fill}
\title{discrete palette for fill aesthetic}
\usage{
discrete_fill()
}
\value{
discrete scale
}
\description{
n <= 9: RColorBrewer Set1\cr 
n > 9: hues "I want hue"
}
\examples{
ggplot2::ggplot(data.frame(x=c("A","B","C"),y=1:3), ggplot2::aes(x=x,y=y,fill=x)) + ggplot2::geom_bar(stat="identity") + discrete_fill()

}

